# [add] changed RedisAI_Run_ThreadMain to only sleep if the last Pop was empty